Adds semantic-ui-react
For projects that use soda-ash, we need the `semanticUIReact` global
that cljsjs provides, instead of just relying directly on an
npm-installed version.

Note that brings in the ES module instead of the UMD like real cljsjs
does. The UMD fails with an (apparently) unsatisfiable dependency on
`ReactDOM` and was packaged with its own lodash, etc.